projects/myproplex/requirements.txt
anthropic>=0.40.0
tavily-python>=0.3.0
python-dotenv>=1.0.0
streamlit>=1.32.0
